Allow eoc recurrance times to be variable, add new portal storm scenario. (#52952)
