+++ data/json/portal_storm_effect_on_condition.json
@@ -24,6 +24,33 @@
     ]
   },
   {
+    "type": "effect_on_condition",
+    "id": "scenario_quick_portal_storm",
+    "eoc_type": "SCENARIO_SPECIFIC",
+    "effect": [
+      {
+        "arithmetic": [
+          { "global_val": "var", "var_name": "min_increase", "type": "weather", "context": "ps_time" },
+          "=",
+          { "time": "15 minutes" }
+        ]
+      },
+      {
+        "arithmetic": [
+          { "global_val": "var", "var_name": "max_increase", "type": "weather", "context": "ps_time" },
+          "=",
+          { "time": "30 minutes" }
+        ]
+      },
+      {
+        "arithmetic": [ { "global_val": "var", "var_name": "min_woc", "type": "weather", "context": "ps_time" }, "=", { "time": "2 days" } ]
+      },
+      {
+        "arithmetic": [ { "global_val": "var", "var_name": "max_woc", "type": "weather", "context": "ps_time" }, "=", { "time": "4 days" } ]
+      }
+    ]
+  },
+  {
     "id": "test_portal_cause",
     "type": "TOOL",
     "name": { "str_sp": "Test Portal Causer" },
@@ -61,8 +88,8 @@
   {
     "type": "effect_on_condition",
     "id": "EOC_PORTAL_STORM_WARN_OR_CAUSE_RECURRING",
-    "recurrence_min": "6 days",
-    "recurrence_max": "9 days",
+    "recurrence_min": { "name": "min_woc", "type": "weather", "context": "ps_time", "default": "6 days", "global": true },
+    "recurrence_max": { "name": "max_woc", "type": "weather", "context": "ps_time", "default": "9 days", "global": true },
     "global": true,
     "effect": [ { "set_queue_effect_on_condition": [ "EOC_PORTAL_STORM_WARN" ] } ]
   },
@@ -261,8 +288,8 @@
     "type": "effect_on_condition",
     "id": "EOC_PORTAL_STRENGTHEN",
     "//": "Increases the portal storms strength.",
-    "recurrence_min": "45 minutes",
-    "recurrence_max": "75 minutes",
+    "recurrence_min": { "name": "min_increase", "type": "weather", "context": "ps_time", "default": "45 minutes", "global": true },
+    "recurrence_max": { "name": "max_increase", "type": "weather", "context": "ps_time", "default": "75 minutes", "global": true },
     "global": true,
     "condition": { "is_weather": "portal_storm" },
     "deactivate_condition": { "not": { "is_weather": "portal_storm" } },
