+++ data/json/scenarios.json
@@ -661,7 +661,7 @@
   {
     "type": "scenario",
     "id": "strong_portal_storm",
-    "name": "Powerful portal storm",
+    "name": "Outside In Strong Portal Storm",
     "points": -3,
     "description": "A powerful portal storm is happening and you need to find shelter.",
     "flags": [ "CHALLENGE", "LONE_START" ],
@@ -672,13 +672,24 @@
   {
     "type": "scenario",
     "id": "strong_portal_storm_base",
-    "name": "Weakened reality",
+    "name": "Stronger Portal Storms",
     "points": -4,
     "description": "You've managed to get to a shelter, but reality has grown weaker than normal and portal storms will be quite strong.",
     "flags": [ "CHALLENGE", "LONE_START" ],
     "start_name": "Evac Shelter",
     "allowed_locs": [ "sloc_shelter_a", "sloc_shelter_b", "sloc_shelter_c" ],
     "eoc": [ "scenario_strong_portal_storm_base" ]
+  },
+  {
+    "type": "scenario",
+    "id": "quick_portal_storm_base",
+    "name": "Frequent Portal Storms",
+    "points": -4,
+    "description": "You've managed to get to a shelter, but reality has grown weaker than normal, portal storms will be more frequent and strengthen faster.",
+    "flags": [ "CHALLENGE", "LONE_START" ],
+    "start_name": "Evac Shelter",
+    "allowed_locs": [ "sloc_shelter_a", "sloc_shelter_b", "sloc_shelter_c" ],
+    "eoc": [ "scenario_quick_portal_storm" ]
   },
   {
     "type": "scenario",
