+++ doc/EFFECT_ON_CONDITION.md
@@ -5,8 +5,8 @@ An effect_on_condition is an object allowing the combination of dialog condition
 
 |Identifier|Type|Description|
 |-|-|-|
-| `recurrence_min`| int | The effect_on_condition is automatically invoked (activated) with at least this many seconds in-between.
-| `recurrence_max`| int | The effect_on_condition is automatically invoked (activated) at least once this many seconds.
+| `recurrence_min`| int or variable object | The effect_on_condition is automatically invoked (activated) with at least this many seconds in-between. If it is an object it must have strings `name`, `type`, and `context`. `default` can be either an int or a string describing a time span. `global` is an optional bool (default false), if it is true the variable used will always be from the player character rather than the target of the dialog.
+| `recurrence_max`| int or variable name | The effect_on_condition is automatically invoked (activated) at least once this many seconds. If it is an object it must have strings `name`, `type`, and `context`. `default` can be either an int or a string describing a time span. `global` is an optional bool (default false), if it is true the variable used will always be from the player character rather than the target of the dialog.
 | `condition`| condition | The condition(s) under which this effect_on_condition, upon activation, will cause its effect.  See the "Dialogue conditions" section of [NPCs](NPCs.md) for the full syntax.
 | `deactivate_condition`| condition | *optional* When an effect_on_condition is automatically activated (invoked) and fails its condition(s), `deactivate_condition` will be tested if it exists and there is no `false_effect` entry.  If it returns true, this effect_on_condition will no longer be invoked automatically every `recurrence_max` seconds.  Whenever the player/npc gains/loses a trait or bionic all deactivated effect_on_conditions will have `deactivate_condition` run; on a return of false, the effect_on_condition will start being run again.  This is to allow adding effect_on_conditions for specific traits or bionics that don't waste time running when you don't have the target bionic/trait.  See the "Dialogue conditions" section of [NPCs](NPCs.md) for the full syntax.
 | `effect`| effect | The effect(s) caused if `condition` returns true upon activation.  See the "Dialogue Effects" section of [NPCs](NPCs.md) for the full syntax.
