+++ src/effect_on_condition.cpp
@@ -62,11 +62,8 @@ void effect_on_condition::load( const JsonObject &jo, const std::string & )
             jo.throw_error( "A recurring effect_on_condition must be of type RECURRING." );
         }
         type = eoc_type::RECURRING;
-        mandatory( jo, was_loaded, "recurrence_min", recurrence_min );
-        mandatory( jo, was_loaded, "recurrence_max", recurrence_max );
-        if( recurrence_max < recurrence_min ) {
-            jo.throw_error( "recurrence_max cannot be smaller than recurrence_min." );
-        }
+        recurrence_min = get_duration_or_var( jo, "recurrence_min", false );
+        recurrence_max = get_duration_or_var( jo, "recurrence_max", false );
     }
     if( type == eoc_type::NUM_EOC_TYPES ) {
         type = eoc_type::ACTIVATION;
@@ -109,9 +106,9 @@ effect_on_condition_id effect_on_conditions::load_inline_eoc( const JsonValue &j
     }
 }
 
-static time_duration next_recurrence( const effect_on_condition_id &eoc )
+static time_duration next_recurrence( const effect_on_condition_id &eoc, talker *talk )
 {
-    return rng( eoc->recurrence_min, eoc->recurrence_max );
+    return rng( eoc->recurrence_min.evaluate( talk ), eoc->recurrence_max.evaluate( talk ) );
 }
 
 void effect_on_conditions::load_new_character( Character &you )
@@ -129,7 +126,7 @@ void effect_on_conditions::load_new_character( Character &you )
 
     for( const effect_on_condition &eoc : effect_on_conditions::get_all() ) {
         if( eoc.type == eoc_type::RECURRING && ( ( is_avatar && eoc.global ) || !eoc.global ) ) {
-            queued_eoc new_eoc = queued_eoc{ eoc.id, calendar::turn + next_recurrence( eoc.id ) };
+            queued_eoc new_eoc = queued_eoc{ eoc.id, calendar::turn + next_recurrence( eoc.id, get_talker_for( you ).get() )};
             if( eoc.global ) {
                 g->queued_global_effect_on_conditions.push( new_eoc );
             } else {
@@ -184,7 +181,8 @@ void effect_on_conditions::load_existing_character( Character &you )
 
     for( const std::pair<const effect_on_condition_id, bool> &eoc_pair : new_eocs ) {
         if( eoc_pair.second ) {
-            queue_effect_on_condition( next_recurrence( eoc_pair.first ), eoc_pair.first, you );
+            queue_effect_on_condition( next_recurrence( eoc_pair.first, get_talker_for( you ).get() ),
+                                       eoc_pair.first, you );
         }
     }
 }
@@ -212,11 +210,11 @@ static void process_eocs( std::priority_queue<queued_eoc, std::vector<queued_eoc
         bool activated = top.eoc->activate( d );
         if( top.eoc->type == eoc_type::RECURRING ) {
             if( activated ) { // It worked so add it back
-                queued_eoc new_eoc = queued_eoc{ top.eoc, calendar::turn + next_recurrence( top.eoc ) };
+                queued_eoc new_eoc = queued_eoc{ top.eoc, calendar::turn + next_recurrence( top.eoc, d.actor( false ) ) };
                 eocs_to_queue.push_back( new_eoc );
             } else {
                 if( !top.eoc->check_deactivate() ) { // It failed but shouldn't be deactivated so add it back
-                    queued_eoc new_eoc = queued_eoc{ top.eoc, calendar::turn + next_recurrence( top.eoc ) };
+                    queued_eoc new_eoc = queued_eoc{ top.eoc, calendar::turn + next_recurrence( top.eoc, d.actor( false ) ) };
                     eocs_to_queue.push_back( new_eoc );
                 } else { // It failed and should be deactivated for now
                     eoc_vector.push_back( top.eoc );
@@ -244,7 +242,7 @@ void effect_on_conditions::process_effect_on_conditions( Character &you )
 static void process_reactivation( std::vector<effect_on_condition_id>
                                   &inactive_effect_on_condition_vector,
                                   std::priority_queue<queued_eoc, std::vector<queued_eoc>, eoc_compare>
-                                  &queued_effect_on_conditions )
+                                  &queued_effect_on_conditions, talker *talk )
 {
     std::vector<effect_on_condition_id> ids_to_reactivate;
     for( const effect_on_condition_id &eoc : inactive_effect_on_condition_vector ) {
@@ -253,7 +251,7 @@ static void process_reactivation( std::vector<effect_on_condition_id>
         }
     }
     for( const effect_on_condition_id &eoc : ids_to_reactivate ) {
-        queued_effect_on_conditions.push( queued_eoc{ eoc, calendar::turn + next_recurrence( eoc ) } );
+        queued_effect_on_conditions.push( queued_eoc{ eoc, calendar::turn + next_recurrence( eoc, talk ) } );
         inactive_effect_on_condition_vector.erase( std::remove(
                     inactive_effect_on_condition_vector.begin(), inactive_effect_on_condition_vector.end(),
                     eoc ), inactive_effect_on_condition_vector.end() );
@@ -262,13 +260,14 @@ static void process_reactivation( std::vector<effect_on_condition_id>
 
 void effect_on_conditions::process_reactivate( Character &you )
 {
-    process_reactivation( you.inactive_effect_on_condition_vector, you.queued_effect_on_conditions );
+    process_reactivation( you.inactive_effect_on_condition_vector, you.queued_effect_on_conditions,
+                          get_talker_for( you ).get() );
 }
 
 void effect_on_conditions::process_reactivate()
 {
     process_reactivation( g->inactive_global_effect_on_condition_vector,
-                          g->queued_global_effect_on_conditions );
+                          g->queued_global_effect_on_conditions, get_talker_for( get_avatar() ).get() );
 }
